Renamed helpe.rs to common.rs
shares basic UI-related helper routines or structures
